requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12
opencv-contrib-python==4.10.0.84
numpy>=1.26,<3.0
Pillow>=10.4
scikit-image>=0.24
scipy>=1.14
pytest>=8.0
httpx>=0.27
